Make changes for use in C++ projects
In order to use the code in a Tensor Flow Op, some basic changes were made to improve comptability with a C++ compiler. Specifically, malloc and calloc calls are properly cast to the desired type.

In addition, this includes a breaking API change, enabling calculating the consensus contour size from either the config or setup. For the config, use `cccSizeConfig` and for the setup, use `cccSizeSetup`.